Relocate DracoHeader to the top-level namespace
